requirements.txt
requests>=2.28.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
